Rearranging text macros
